Commit 21 - all modules completed and ready to check

+++ cut/src/reader/reader.h
@@ -17,7 +17,7 @@
 typedef struct reader Reader;
 
 // PROTOTYPE FUNCTIONS FOR OUTSIDE WORLD
-Reader* Reader_init(Buffer* const, const long); 
+Reader* Reader_init(Buffer* const, const uint8_t); 
 int Reader_start(Reader* const, volatile sig_atomic_t*); 
 int Reader_join(Reader* const);
 void Reader_destroy(Reader*);
